# Optimize `Nerves.Runtime.Heart.parse_cmd/1`