@fhunleth 2022-11-22
You really like `for` comprehensions. I admire your passion for this particular language feature.